requirements.txt
fastapi
uvicorn
requests
python-dotenv
pip freeze > requirements.txt

